Update and rename u_files.js to u_move_files.js
if --create is set to true, then this file makes directory for each unique file extension on --target if --ext is not set to specific extension. if --ext is set to extension, it will only make a directory for the extension.
if --delete is set to true, then this file deletes all directories on --target if --ext is not set to specific extension. if --ext is set to extension, it will only delete the directory for that extension.